[FLAMBE-18]ユーザに紐づくタスクを参照できるようにする feat: ユーザの作業中タスクデータを取得できるように修正
